% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_packages.R
\name{add_packages}
\alias{add_packages}
\title{Compile Packages and Add to CRAN Repository}
\usage{
add_packages(
  packages,
  repository,
  available_packages = available.packages(),
  quiet = TRUE
)
}
\arguments{
\item{packages}{A character vector of package names or package directories}

\item{repository}{A file path to a CRAN repository.}

\item{available_packages}{An object like the one generated by
\code{available.packages()}}

\item{quiet}{boolean, TRUE silences output.}
}
\description{
For a vector of package names and package directories, a path to a directory
containing a CRAN repository
}
