% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_packages.R
\name{available_packages-param}
\alias{available_packages-param}
\title{Available Packages}
\arguments{
\item{available_packages}{An object like the one generated by
\code{available.packages()}}
}
\description{
Available Packages
}
