% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{create_repository}
\alias{create_repository}
\title{Create a bare repository}
\usage{
create_repository(repo_dir)
}
\arguments{
\item{repo_dir}{file path to a repository}
}
\value{
The original repo_dir argument
}
\description{
Create a bare repository
}
